Add AppUserService
Add functionality to pull an app_user entry from the database. This will
allow future functionality such as obtaining a user's password salt for
client authentication and pulling the user's password hash to
authenticate them.

Add functionality to throw an exception when there is no entry for a
user with the provided username in the database.