Update KerberosRestController
Update KerberosRestController to handle any exceptions caused by a
malformed TicketGrantingTicket object and to ensure both the tgt and the
session authenticator sent to the user contain the same session key.
This session key was made to be a base64 string for interoperability
between Spring and AngularJS.

As of this point, there is no encryption for this process or checking of
userroles.