Update LoginService
Update login service to provide functionality for the entire
authentication process from getting a user's salt to sending the
authenticator and recieving a TGT. Currently no encryption is applied to
this functionality.

Update passHash method to return a promise as to not freeze the user's
browser when computing the costly hash operation.